refactor: 프로젝트 전체 폴더 구조 변경 및 경로 변경
