+++ src/pages/EvaluationPage.js
@@ -1,5 +1,5 @@
 import { styled } from "styled-components";
-import MainBg from "../MainBg";
+import MainBg from "../components/MainBg";
 import { Container } from "react-bootstrap";
 import { useState } from "react";
 
@@ -18,12 +18,11 @@ const TheadStyle = styled.thead`
   font-weight : bold;
 `;
 
-function CommunityPage(){
+function EvaluationPage(){
   let [num, setNum ] = useState([0,0,0,0,0]);
-
   return(
     <>
-      <MainBg title="Community" subtitle="우리들의 대화공간"/>
+      <MainBg title="Evaluation Index" subtitle="기업 평가 지표"/>
 
       <Container>
         <h3 className="mb-5">ESG 기업 정보</h3>
@@ -31,9 +30,9 @@ function CommunityPage(){
           <TheadStyle>
             <tr>
               <th scope="col">#</th>
-              <th scope="col">제목</th>
-              <th scope="col">작성자</th>
-              <th scope="col">조회수</th>
+              <th scope="col">기업 이름</th>
+              <th scope="col">평가 등급</th>
+              <th scope="col">추천</th>
             </tr>
           </TheadStyle>
           <tbody>
@@ -62,4 +61,4 @@ function CommunityPage(){
   )
 };
 
-export default CommunityPage;
+export default EvaluationPage;
